Revert "Nullable type (#204)"
This reverts commit 692cab5651b47b6a3a9ea2b7930848451dd89d85.
