+++ sdk/core/azure-core/test/CMakeLists.txt
@@ -11,9 +11,8 @@ set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 add_executable (
      ${TARGET_NAME}
-     main.cpp
-     nullable.cpp
-     http.cpp)
+     main.cpp 
+)
 
 target_link_libraries(${TARGET_NAME} PRIVATE azure-core)
 add_gtest(${TARGET_NAME})
